errors fixed

+++ public/components/resource.js
@@ -17,8 +17,6 @@ var alltags = document.querySelectorAll('.alltags');
 // storing in local storage variables;
 let resources = [];
 
-let tagColours = ['#B7B893', '#6E703D', '#B2A496'];
-
 class ResourceObj {
     constructor(data, summary, tags) {
         this.data = data;
@@ -73,6 +71,7 @@ function generatePreview(data) {
 
     let sendTag = document.createElement('button');
     sendTag.setAttribute('id', 'sendTagBtn');
+    sendTag.setAttribute('type', "submit");
 
     let sendIcon = document.createElement('i');
     sendIcon.setAttribute('class', 'fa fa-paper-plane-o');
@@ -138,7 +137,7 @@ function generatePreview(data) {
         ls.push(reObj);
         localStorage.setItem('resources', JSON.stringify(ls));
     }
-
+    window.location.reload();
 }
 
 // Retrieve resource data from local storage, through going through the list and check whether summary is inputed. 
@@ -177,6 +176,7 @@ function retreiveLocalRes() {
             }
 
             let sendTag = document.createElement('button');
+            sendTag.setAttribute('type', "submit");
             sendTag.setAttribute('id', 'sendTagBtn');
 
             let sendIcon = document.createElement('i');
@@ -261,8 +261,6 @@ function assignTag() {
     tagResInput = document.querySelectorAll('#tagResInput');
     sendTagBtn = document.querySelectorAll('#sendTagBtn');
 
-    console.log(leftCols);
-
     let list = localStorage.getItem('resources');
     list = JSON.parse(list);
 
@@ -270,20 +268,30 @@ function assignTag() {
 
     for (let i = 0; i < leftCols.length; i++) {
         sendTagBtn[i].addEventListener('click', function (i) {
+            var exist = false;
             cursorLs = list[i].tags;
             alltags = leftCols[i].querySelector('.alltags');
             if (tagResInput[i].value != '') {
-                cursorLs.push(tagResInput[i].value);
-                let resTag = document.createElement('button');
-                resTag.setAttribute('id', 'resTag');
-                resTag.innerHTML = tagResInput[i].value;
-                alltags.appendChild(resTag);
+                for (let j = 0; j < cursorLs.length; j++) {
+                    if (tagResInput[i].value == cursorLs[j]) {
+                        exist = true;
+                        alert('Tag already exist, give it another name.');
+                        break;
+                    }
+                }
+                if (exist == false) {
+                    cursorLs.push(tagResInput[i].value);
+                    let resTag = document.createElement('button');
+                    resTag.setAttribute('id', 'resTag');
+                    resTag.innerHTML = tagResInput[i].value;
+                    alltags.appendChild(resTag);
 
-                tagResInput[i].value = '';
+                    tagResInput[i].value = '';
 
-                // Put updated tag ls into local storage;
-                list[i].tags = cursorLs;
-                localStorage.setItem('resources', JSON.stringify(list));
+                    // Put updated tag ls into local storage;
+                    list[i].tags = cursorLs;
+                    localStorage.setItem('resources', JSON.stringify(list));
+                }
             }
         }.bind(null, i));
     }
